+++ modules/structs/settings.go
@@ -9,3 +9,8 @@ type GeneralRepoSettings struct {
 	MirrorsDisabled bool `json:"mirrors_disabled"`
 	HTTPGitDisabled bool `json:"http_git_disabled"`
 }
+
+// GeneralUISettings contains global ui settings exposed by API
+type GeneralUISettings struct {
+	AllowedReactions []string `json:"allowed_reactions"`
+}
