+++ routers/api/v1/settings/settings.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-package misc
+package settings
 
 import (
 	"net/http"
@@ -12,22 +12,24 @@ import (
 	api "code.gitea.io/gitea/modules/structs"
 )
 
-// SettingGetsAllowedReactions return allowed reactions
-func SettingGetsAllowedReactions(ctx *context.APIContext) {
-	// swagger:operation GET /settings/allowed_reactions miscellaneous getAllowedReactions
+// GetGeneralUISettings returns instance's global settings for ui
+func GetGeneralUISettings(ctx *context.APIContext) {
+	// swagger:operation GET /settings/ui settings getGeneralUISettings
 	// ---
-	// summary: Returns string array of allowed reactions
+	// summary: Get instance's global settings for ui
 	// produces:
 	// - application/json
 	// responses:
 	//   "200":
-	//     "$ref": "#/responses/StringSlice"
-	ctx.JSON(http.StatusOK, setting.UI.Reactions)
+	//     "$ref": "#/responses/GeneralUISettings"
+	ctx.JSON(http.StatusOK, api.GeneralUISettings{
+		AllowedReactions: setting.UI.Reactions,
+	})
 }
 
 // GetGeneralRepoSettings returns instance's global settings for repositories
 func GetGeneralRepoSettings(ctx *context.APIContext) {
-	// swagger:operation GET /settings/repository miscellaneous getGeneralRepositorySettings
+	// swagger:operation GET /settings/repository settings getGeneralRepositorySettings
 	// ---
 	// summary: Get instance's global settings for repositories
 	// produces:
