+++ routers/api/v1/swagger/misc.go
@@ -21,10 +21,3 @@ type swaggerResponseStringSlice struct {
 	// in:body
 	Body []string `json:"body"`
 }
-
-// GeneralRepoSettings
-// swagger:response GeneralRepoSettings
-type swaggerResponseGeneralRepoSettings struct {
-	// in:body
-	Body api.GeneralRepoSettings `json:"body"`
-}
